+++ src/main/java/de/gematik/rbellogger/converter/RbelConverterPlugin.java
@@ -21,4 +21,8 @@ import de.gematik.rbellogger.data.RbelElement;
 public interface RbelConverterPlugin {
 
     void consumeElement(RbelElement rbelElement, RbelConverter converter);
+
+    default boolean ignoreOversize() {
+        return false;
+    }
 }