+++ src/main/java/de/gematik/rbellogger/data/facet/RbelBearerTokenFacet.java
@@ -47,7 +47,7 @@ public class RbelBearerTokenFacet implements RbelFacet {
             public ContainerTag performRendering(final RbelElement element, final Optional<String> key,
                                                  final RbelHtmlRenderingToolkit renderingToolkit) {
                 return div(t1ms("Bearer Token")
-                    .with(showContentButtonAndDialog(element)))
+                    .with(showContentButtonAndDialog(element, renderingToolkit)))
                     .with(addNotes(element, "mb-5"))
                     .with(ancestorTitle().with(
                         vertParentTitle().with(