+++ src/main/java/de/gematik/rbellogger/data/facet/RbelJweFacet.java
@@ -49,7 +49,7 @@ public class RbelJweFacet implements RbelFacet {
             @Override
             public ContainerTag performRendering(RbelElement element, Optional<String> key,
                                                  RbelHtmlRenderingToolkit renderingToolkit) {
-                return div(t1ms("JWE").with(showContentButtonAndDialog(element)))
+                return div(t1ms("JWE").with(showContentButtonAndDialog(element, renderingToolkit)))
                     .with(addNotes(element, "mb-5"))
                     .with(ancestorTitle().with(
                             vertParentTitle().with(